fix cluster

+++ nanopypes/compute.py
@@ -84,7 +84,7 @@ class NanopypesCluster:
     def _build_lsf(self):
         ncpus = int((self.num_workers / self.workers_per_job)) * self.worker_cores
         mem_bytes = self.worker_memory * self.workers_per_job * 1024**2
-        dask_memory = str(self.worker_memory / 1024) + 'GB'
+        dask_memory = str(int(self.worker_memory / 1024)) + 'GB'
         cluster = LSFCluster(queue=self.queue,  # Passed to #BSUB -q option.
                                   project=self.project,  # Passed to #BSUB -P option.
                                   processes=self.workers_per_job,
